MAINT: Moving calibration class to sndmotor since it can be used by regular motors
